Make ProjectTeamTable component to reuse in project create stepper
